dev_requirements.txt
dependencymanager==1.0.0
autopep8==1.5.5
toml==0.10.2
pycodestyle==2.6.0
twine==3.3.0
requests==2.25.1
urllib3==1.26.3
idna==2.10
certifi==2020.12.5
chardet==4.0.0
rfc3986==1.4.0
pkginfo==1.7.0
tqdm==4.56.2
readme-renderer==28.0
docutils==0.16
pygments==2.8.0
six==1.15.0
bleach==3.3.0
packaging==20.9
pyparsing==2.4.7
webencodings==0.5.1
keyring==22.0.1
jeepney==0.6.0
secretstorage==3.3.1
cryptography==3.4.6
cffi==1.14.5
pycparser==2.20
colorama==0.4.4
requests-toolbelt==0.9.1
PyYAML==5.4.1